refactor: Reorganize the project structure
